refactor: remove AuthProvider component and update Supabase client imports for improved structure and clarity
